+++ src/Mpociot/BotMan/Conversation.php
@@ -4,6 +4,7 @@ namespace Mpociot\BotMan;
 
 use Closure;
 use Illuminate\Support\Collection;
+use Mpociot\BotMan\Messages\Matcher;
 use Mpociot\BotMan\Interfaces\ShouldQueue;
 
 /**
@@ -53,7 +54,7 @@ abstract class Conversation
     public function askForImages($question, $next, $repeat = null, $additionalParameters = [])
     {
         $additionalParameters['__getter'] = 'getImages';
-        $additionalParameters['__pattern'] = BotMan::IMAGE_PATTERN;
+        $additionalParameters['__pattern'] = Matcher::IMAGE_PATTERN;
         $additionalParameters['__repeat'] = ! is_null($repeat) ? $this->bot->serializeClosure($repeat) : $repeat;
 
         return $this->ask($question, $next, $additionalParameters);
@@ -69,7 +70,7 @@ abstract class Conversation
     public function askForVideos($question, $next, $repeat = null, $additionalParameters = [])
     {
         $additionalParameters['__getter'] = 'getVideos';
-        $additionalParameters['__pattern'] = BotMan::VIDEO_PATTERN;
+        $additionalParameters['__pattern'] = Matcher::VIDEO_PATTERN;
         $additionalParameters['__repeat'] = ! is_null($repeat) ? $this->bot->serializeClosure($repeat) : $repeat;
 
         return $this->ask($question, $next, $additionalParameters);
@@ -85,7 +86,7 @@ abstract class Conversation
     public function askForAudio($question, $next, $repeat = null, $additionalParameters = [])
     {
         $additionalParameters['__getter'] = 'getAudio';
-        $additionalParameters['__pattern'] = BotMan::AUDIO_PATTERN;
+        $additionalParameters['__pattern'] = Matcher::AUDIO_PATTERN;
         $additionalParameters['__repeat'] = ! is_null($repeat) ? $this->bot->serializeClosure($repeat) : $repeat;
 
         return $this->ask($question, $next, $additionalParameters);
@@ -101,7 +102,7 @@ abstract class Conversation
     public function askForLocation($question, $next, $repeat = null, $additionalParameters = [])
     {
         $additionalParameters['__getter'] = 'getLocation';
-        $additionalParameters['__pattern'] = BotMan::LOCATION_PATTERN;
+        $additionalParameters['__pattern'] = Matcher::LOCATION_PATTERN;
         $additionalParameters['__repeat'] = ! is_null($repeat) ? $this->bot->serializeClosure($repeat) : $repeat;
 
         return $this->ask($question, $next, $additionalParameters);