+++ src/Mpociot/BotMan/Drivers/FacebookLocationDriver.php
@@ -2,9 +2,9 @@
 
 namespace Mpociot\BotMan\Drivers;
 
-use Mpociot\BotMan\BotMan;
 use Mpociot\BotMan\Message;
 use Illuminate\Support\Collection;
+use Mpociot\BotMan\Messages\Matcher;
 use Mpociot\BotMan\Attachments\Location;
 
 class FacebookLocationDriver extends FacebookDriver
@@ -42,7 +42,7 @@ class FacebookLocationDriver extends FacebookDriver
         $messages = Collection::make($this->event->get('messaging'))->filter(function ($msg) {
             return isset($msg['message']) && isset($msg['message']['attachments']) && isset($msg['message']['attachments']);
         })->transform(function ($msg) {
-            $message = new Message(BotMan::LOCATION_PATTERN, $msg['recipient']['id'], $msg['sender']['id'], $msg);
+            $message = new Message(Matcher::LOCATION_PATTERN, $msg['recipient']['id'], $msg['sender']['id'], $msg);
             $message->setLocation($this->getLocation($msg));
 
             return $message;