+++ src/Mpociot/BotMan/Drivers/SlackRTMDriver.php
@@ -179,7 +179,9 @@ class SlackRTMDriver implements DriverInterface
      */
     public function replyInThread($message, $additionalParameters, $matchingMessage)
     {
-        $additionalParameters['thread_ts'] = $matchingMessage->getPayload()->get('ts');
+        $additionalParameters['thread_ts'] = ! empty($matchingMessage->getPayload()->get('thread_ts'))
+            ? $matchingMessage->getPayload()->get('thread_ts')
+            : $matchingMessage->getPayload()->get('ts');
 
         return $this->reply($message, $matchingMessage, $additionalParameters);
     }