+++ src/Mpociot/BotMan/Drivers/TelegramVideoDriver.php
@@ -3,8 +3,8 @@
 namespace Mpociot\BotMan\Drivers;
 
 use Mpociot\BotMan\Attachments\Video;
-use Mpociot\BotMan\BotMan;
 use Mpociot\BotMan\Message;
+use Mpociot\BotMan\Messages\Matcher;
 
 class TelegramVideoDriver extends TelegramDriver
 {
@@ -27,7 +27,7 @@ class TelegramVideoDriver extends TelegramDriver
      */
     public function getMessages()
     {
-        $message = new Message(BotMan::VIDEO_PATTERN, $this->event->get('from')['id'], $this->event->get('chat')['id'], $this->event);
+        $message = new Message(Matcher::VIDEO_PATTERN, $this->event->get('from')['id'], $this->event->get('chat')['id'], $this->event);
         $message->setVideos($this->getVideos());
 
         return [$message];