+++ src/Mpociot/BotMan/Drivers/WeChatPhotoDriver.php
@@ -3,8 +3,8 @@
 namespace Mpociot\BotMan\Drivers;
 
 use Mpociot\BotMan\Attachments\Image;
-use Mpociot\BotMan\BotMan;
 use Mpociot\BotMan\Message;
+use Mpociot\BotMan\Messages\Matcher;
 
 class WeChatPhotoDriver extends WeChatDriver
 {
@@ -37,7 +37,7 @@ class WeChatPhotoDriver extends WeChatDriver
      */
     public function getMessages()
     {
-        $message = new Message(BotMan::IMAGE_PATTERN, $this->event->get('ToUserName'), $this->event->get('FromUserName'), $this->event);
+        $message = new Message(Matcher::IMAGE_PATTERN, $this->event->get('ToUserName'), $this->event->get('FromUserName'), $this->event);
         $message->setImages($this->getImages());
 
         return [$message];