+++ src/Mpociot/BotMan/Drivers/WeChatVideoDriver.php
@@ -3,8 +3,8 @@
 namespace Mpociot\BotMan\Drivers;
 
 use Mpociot\BotMan\Attachments\Video;
-use Mpociot\BotMan\BotMan;
 use Mpociot\BotMan\Message;
+use Mpociot\BotMan\Messages\Matcher;
 
 class WeChatVideoDriver extends WeChatDriver
 {
@@ -34,7 +34,7 @@ class WeChatVideoDriver extends WeChatDriver
      */
     public function getMessages()
     {
-        $message = new Message(BotMan::VIDEO_PATTERN, $this->event->get('ToUserName'),
+        $message = new Message(Matcher::VIDEO_PATTERN, $this->event->get('ToUserName'),
             $this->event->get('FromUserName'), $this->event);
         $message->setVideos($this->getVideo());