+++ tests/Drivers/TelegramPhotoDriverTest.php
@@ -3,9 +3,9 @@
 namespace Mpociot\BotMan\Tests\Drivers;
 
 use Mockery as m;
-use Mpociot\BotMan\BotMan;
 use Mpociot\BotMan\Http\Curl;
 use PHPUnit_Framework_TestCase;
+use Mpociot\BotMan\Messages\Matcher;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Mpociot\BotMan\Drivers\TelegramPhotoDriver;
@@ -128,7 +128,7 @@ class TelegramPhotoDriverTest extends PHPUnit_Framework_TestCase
             ],
         ], $htmlInterface);
         $message = $driver->getMessages()[0];
-        $this->assertSame(BotMan::IMAGE_PATTERN, $message->getText());
+        $this->assertSame(Matcher::IMAGE_PATTERN, $message->getText());
         $this->assertSame('https://api.telegram.org/file/bot/foo', $message->getImages()[0]->getUrl());
         $this->assertSame([
 	        'file_id' => 'AgADAgAD6KcxG4tSUUnK3tsu3YsxCu8VSw0ABO72aPxtHuGxcGMFAAEC',